+++ app/src/main/java/io/github/froodyapp/ui/EntryTypeSelectionDialog.java
@@ -12,9 +12,11 @@ import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.TextView;
 
+import java.text.Collator;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.Locale;
 import java.util.Random;
 
 import butterknife.BindView;
@@ -107,9 +109,11 @@ public class EntryTypeSelectionDialog extends DialogFragment implements EntryTyp
         for (int i = FroodyEntryFormatter.ENTRY_TYPE_MIN; i < entryTypeResource.length; i++) {
             list.add(i);
         }
+        final Collator collator = Collator.getInstance(getResources().getConfiguration().locale);
+        collator.setStrength(Collator.SECONDARY);
         Collections.sort(list, new Comparator<Integer>() {
             public int compare(final Integer o1, final Integer o2) {
-                return entryTypeResource[o1].compareTo(entryTypeResource[o2]);
+                return collator.compare(entryTypeResource[o1],entryTypeResource[o2]);
             }
         });
         list.add(0, FroodyEntryFormatter.ENTRY_TYPE_CUSTOM);